Lyric Locate V2
This is a functional and very polished version of our code. When it is finished, it should run as a .py, it should explain anything else that needs to be explained, and it should be ideally error-proof. I would like the final version to be defined as Lyric_Locate() so its your job to do that